Chunk is now marked as dirty; saving only dirty chunks; rewritten load / save not to use cChunkPtr; set VC2008 project to level4 warnings; block entities are now loaded and saved properly
git-svn-id: http://mc-server.googlecode.com/svn/trunk@273 0a769ca7-a7f5-676a-18bf-c427514a06d6

+++ source/cPlayer.cpp
@@ -922,7 +922,7 @@ bool cPlayer::SaveToDisk()
 		LOGERROR("ERROR WRITING PLAYER \"%s\" TO FILE \"%s\" - cannot open file", m_pState->PlayerName.c_str(), SourceFile.c_str());
 		return false;
 	}
-	if (f.Write(JsonData.c_str(), JsonData.size()) != JsonData.size())
+	if (f.Write(JsonData.c_str(), JsonData.size()) != (int)JsonData.size())
 	{
 		LOGERROR("ERROR WRITING PLAYER JSON TO FILE \"%s\"", SourceFile.c_str()); 
 		return false;
